requirements_dev.txt
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
psycopg-binary==3.2.3
python-dotenv==1.0.1
uuid==1.30
